1Lot=1通貨など1Lotのポジションサイズ小さい時、損益が小数になるので、Moneyをlong longからdoubleに変更

+++ FX Simulator/Common.h
@@ -10,7 +10,7 @@
 
 typedef double rate_t;
 
-typedef long long amount_t;
+typedef double amount_t;
 
 typedef unsigned long long lot_t;
